// app/src/api/trip/TripClient.ts
export interface Expense {
  id: string;
  payer: string;
  amount: number;
  description: string;
  split_among: string[];
  created_at: number;
}